+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ﻿## 1.0.4
 
 1.  iOS: Rename `IosPlist.PListSetBoolean` to `IosPlist.SetBoolean`
-2.  Android: Add `AndroidAppManifestBuild.SetApplicationAttribute`
+2.  Android: Rename `AndroidAppManifestBuild` to `AndroidManifest` 
+3.  Android: Add `AndroidManifest.SetApplicationAttribute`
+4.  Android: Add `AndroidManifest.SetActivityWithLauncherIntentAttribute`